+++ src/Geralt/Interop/Interop.Padding.cs
@@ -5,9 +5,9 @@ internal static partial class Interop
     internal static partial class Libsodium
     {
         [DllImport(DllName, CallingConvention = Convention)]
-        internal static extern unsafe int sodium_pad(out int paddedBufferLength, byte* buffer, int unpaddedBufferLength, int blockSize, int maxBufferLength);
+        internal static extern unsafe int sodium_pad(out nuint paddedBufferLength, byte* buffer, nuint unpaddedBufferLength, nuint blockSize, nuint maxBufferLength);
         
         [DllImport(DllName, CallingConvention = Convention)]
-        internal static extern unsafe int sodium_unpad(out int unpaddedBufferLength, byte* paddedBuffer, int paddedBufferLength, int blockSize);
+        internal static extern unsafe int sodium_unpad(out nuint unpaddedBufferLength, byte* paddedBuffer, nuint paddedBufferLength, nuint blockSize);
     }
 }
