+++ src/Geralt/Interop/Interop.SecureRandom.cs
@@ -7,12 +7,12 @@ internal static partial class Interop
         internal const int randombytes_SEEDBYTES = 32;
     
         [DllImport(DllName, CallingConvention = Convention)]
-        internal static extern unsafe void randombytes_buf(byte* buffer, int size);
+        internal static extern unsafe void randombytes_buf(byte* buffer, nuint size);
 
         [DllImport(DllName, CallingConvention = Convention)]
-        internal static extern unsafe void randombytes_buf_deterministic(byte* buffer, int size, byte* seed);
+        internal static extern unsafe void randombytes_buf_deterministic(byte* buffer, nuint size, byte* seed);
 
         [DllImport(DllName, CallingConvention = Convention)]
-        internal static extern int randombytes_uniform(int upperBound);
+        internal static extern int randombytes_uniform(uint upperBound);
     }
 }
